move pipeline into top level
